Refactor to use a facade
