+++ app/controllers/viewing_parties_controller.rb
@@ -1,8 +1,6 @@
 class ViewingPartiesController < ApplicationController
   def new
-    @user = User.find(params[:user_id])
-    @movie = MovieDetailsSearch.new.movie(params[:movie_id])
-    @users = User.where.not(id: params[:user_id])
+    @facade = MovieDetailsFacade.new(params[:user_id], params[:movie_id])
   end
 
   def create
