+++ app/controllers/welcome_controller.rb
@@ -2,6 +2,4 @@ class WelcomeController < ApplicationController
   def index
    @users = User.all
   end
-
-  
 end
