+++ app/facades/movies_search.rb
