--- app/poros/config_search.rb
@@ -1,13 +1,0 @@
-
-require "./app/services/tmdb_service"
-
-class ConfigSearch
-  def base_path
-    data = service.config_base_path
-    data[:images][:base_url]
-  end
-
-  def service 
-    TmdbService.new
-  end
-end
