+++ app/views/users/show.html.erb
@@ -18,7 +18,7 @@ div.party-card {
 <hr>
 <% @viewing_parties.each do |party| %>
   <div class="party-card" id="viewing-party-<%= party.id %>">
-    <img class="movie-poster" src=<%= party.image %>><br>
+    <img class="movie-poster" src=<%= party.image(@user.id) %>><br>
     <%= link_to party.movie_title, user_movie_path(@user, party.movie_id) %>
     <br><br>
     <%= party.date.strftime("%B %d, %Y") %>
