+++ spec/fixtures/vcr_cassettes/Discover_Movies_Page/As_a_user_when_I_visit_users/_id/discover/has_a_search_field_to_enter_movie_title_keywords.yml
@@ -25,7 +25,7 @@ http_interactions:
       Connection:
       - keep-alive
       Date:
-      - Thu, 01 Dec 2022 23:00:55 GMT
+      - Fri, 02 Dec 2022 15:40:09 GMT
       Server:
       - openresty
       Cache-Control:
@@ -36,13 +36,13 @@ http_interactions:
       X-Cache:
       - Miss from cloudfront
       Via:
-      - 1.1 c8997576b0e59159a991d905466fd2d4.cloudfront.net (CloudFront)
+      - 1.1 16d72e0791ff01fc4470d27fc024527a.cloudfront.net (CloudFront)
       X-Amz-Cf-Pop:
-      - HIO50-C2
+      - DEN52-P2
       Alt-Svc:
       - h3=":443"; ma=86400
       X-Amz-Cf-Id:
-      - 3_y470UU3U2UD6Vi4dUu2D357fPuxZHjDOleo3XnmPQEVMbLh3dVEg==
+      - I9phTZfXqwuCr_-T-RaH7FHfbbq8KHCHTfzHHIDrXP7N8OaB4WDohw==
     body:
       encoding: ASCII-8BIT
       string: '{"page":1,"results":[{"adult":false,"backdrop_path":"/kXfqcdQKsToO0OUXHcrrNCHDBzO.jpg","genre_ids":[18,80],"id":278,"original_language":"en","original_title":"The
@@ -51,17 +51,17 @@ http_interactions:
         at the Shawshank prison, where he puts his accounting skills to work for an
         amoral warden. During his long stretch in prison, Dufresne comes to be admired
         by the other inmates -- including an older prisoner named Red -- for his integrity
-        and unquenchable sense of hope.","popularity":81.472,"poster_path":"/hBcY0fE9pfXzvVaY4GKarweriG2.jpg","release_date":"1994-09-23","title":"The
-        Shawshank Redemption","video":false,"vote_average":8.7,"vote_count":22733},{"adult":false,"backdrop_path":"/u6hgtSIzW8nDKKWmun9JDhZz2fp.jpg","genre_ids":[99],"id":546312,"original_language":"en","original_title":"Shawshank:
+        and unquenchable sense of hope.","popularity":80.871,"poster_path":"/hBcY0fE9pfXzvVaY4GKarweriG2.jpg","release_date":"1994-09-23","title":"The
+        Shawshank Redemption","video":false,"vote_average":8.7,"vote_count":22740},{"adult":false,"backdrop_path":"/u6hgtSIzW8nDKKWmun9JDhZz2fp.jpg","genre_ids":[99],"id":546312,"original_language":"en","original_title":"Shawshank:
         The Redeeming Feature","overview":"Film critic and presenter Mark Kermode
         explores the fascination and praise audiences, fans and critics have behind
         the classic The Shawshank Redemption (1994) and how its importance grew higher
         over the years despite being an overlooked film when it was released and not
         getting any Oscars victories when it was nominated as one of the Best Pictures
-        of the Year.","popularity":1.303,"poster_path":"/b7DGM74MRfiLqEveVKwh0jnQ8zX.jpg","release_date":"2001-09-08","title":"Shawshank:
+        of the Year.","popularity":1.366,"poster_path":"/b7DGM74MRfiLqEveVKwh0jnQ8zX.jpg","release_date":"2001-09-08","title":"Shawshank:
         The Redeeming Feature","video":false,"vote_average":8.7,"vote_count":14},{"adult":false,"backdrop_path":null,"genre_ids":[],"id":897459,"original_language":"en","original_title":"Hope
         Springs Eternal: A Look Back at The Shawshank Redemption","overview":"Documentary
-        about the movie.","popularity":0.604,"poster_path":"/vdVab7yNvgYEMd8shCfy2D6nTMu.jpg","release_date":"2004-06-25","title":"Hope
+        about the movie.","popularity":1.4,"poster_path":"/vdVab7yNvgYEMd8shCfy2D6nTMu.jpg","release_date":"2004-06-25","title":"Hope
         Springs Eternal: A Look Back at The Shawshank Redemption","video":false,"vote_average":6.5,"vote_count":2}],"total_pages":1,"total_results":3}'
-  recorded_at: Thu, 01 Dec 2022 23:00:55 GMT
+  recorded_at: Fri, 02 Dec 2022 15:40:09 GMT
 recorded_with: VCR 6.1.0
