+++ spec/fixtures/vcr_cassettes/ViewingParty/instance_methods/_image/returns_the_source_path_for_the_image_related_to_the_movie_party.yml
@@ -2,6 +2,57 @@
 http_interactions:
 - request:
     method: get
+    uri: https://api.themoviedb.org/3/configuration?api_key=<tmdb_key>
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      User-Agent:
+      - Faraday v2.7.1
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+      Accept:
+      - "*/*"
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Content-Type:
+      - application/json;charset=utf-8
+      Transfer-Encoding:
+      - chunked
+      Connection:
+      - keep-alive
+      Date:
+      - Fri, 02 Dec 2022 14:55:37 GMT
+      Server:
+      - openresty
+      Cache-Control:
+      - public, max-age=43200
+      Etag:
+      - W/"8ee6fd095d2be6e301251d766501a916"
+      Vary:
+      - Accept-Encoding,Accept-Encoding,Accept-Encoding
+      - Origin
+      X-Cache:
+      - Hit from cloudfront
+      Via:
+      - 1.1 422414d66021e4e123bcb2b5087f7042.cloudfront.net (CloudFront)
+      X-Amz-Cf-Pop:
+      - DEN52-P2
+      Alt-Svc:
+      - h3=":443"; ma=86400
+      X-Amz-Cf-Id:
+      - 4NgIluW55diPN19zxTzqe91MZUPkipAWU4w3bvZPC89z7kqfJkY7Hw==
+      Age:
+      - '2679'
+    body:
+      encoding: ASCII-8BIT
+      string: '{"images":{"base_url":"http://image.tmdb.org/t/p/","secure_base_url":"https://image.tmdb.org/t/p/","backdrop_sizes":["w300","w780","w1280","original"],"logo_sizes":["w45","w92","w154","w185","w300","w500","original"],"poster_sizes":["w92","w154","w185","w342","w500","w780","original"],"profile_sizes":["w45","w185","h632","original"],"still_sizes":["w92","w185","w300","original"]},"change_keys":["adult","air_date","also_known_as","alternative_titles","biography","birthday","budget","cast","certifications","character_names","created_by","crew","deathday","episode","episode_number","episode_run_time","freebase_id","freebase_mid","general","genres","guest_stars","homepage","images","imdb_id","languages","name","network","origin_country","original_name","original_title","overview","parts","place_of_birth","plot_keywords","production_code","production_companies","production_countries","releases","revenue","runtime","season","season_number","season_regular","spoken_languages","status","tagline","title","translations","tvdb_id","tvrage_id","type","video","videos"]}'
+  recorded_at: Fri, 02 Dec 2022 15:40:16 GMT
+- request:
+    method: get
     uri: https://api.themoviedb.org/3/movie/278?api_key=<tmdb_key>&language=en-US
     body:
       encoding: US-ASCII
@@ -25,7 +76,7 @@ http_interactions:
       Connection:
       - keep-alive
       Date:
-      - Thu, 01 Dec 2022 23:00:58 GMT
+      - Fri, 02 Dec 2022 15:16:22 GMT
       Server:
       - openresty
       Cache-Control:
@@ -35,24 +86,26 @@ http_interactions:
       X-Memc-Key:
       - c7d0f1a47010fc55c00fcec4d5c4a5ef91fba03f
       X-Memc-Age:
-      - '15084'
+      - '6850'
       X-Memc-Expires:
-      - '419'
+      - '7706'
       Etag:
-      - W/"75363ec2263580fdd4ad4795cf87fac1"
+      - W/"f2bb38e62a09ebb696eb76aa63b93a65"
       Vary:
       - Accept-Encoding,Accept-Encoding,Accept-Encoding
       - Origin
       X-Cache:
-      - Miss from cloudfront
+      - Hit from cloudfront
       Via:
-      - 1.1 c3270a6ec9acbf35e35962d1054f78d4.cloudfront.net (CloudFront)
+      - 1.1 8a50fe9452625079391cf0ce7a3e0c56.cloudfront.net (CloudFront)
       X-Amz-Cf-Pop:
-      - HIO50-C2
+      - DEN52-P2
       Alt-Svc:
       - h3=":443"; ma=86400
       X-Amz-Cf-Id:
-      - 8dlfS65rDwE_2RKTB8zwPS2NYb5e_J0KhVUYfsRAvthudgPD7PBdSw==
+      - ve9aOaUDpxOXbKc3xo--19FjO_ByvkNjTYk1OKzpRjtyuwCEXad45w==
+      Age:
+      - '1435'
     body:
       encoding: ASCII-8BIT
       string: '{"adult":false,"backdrop_path":"/kXfqcdQKsToO0OUXHcrrNCHDBzO.jpg","belongs_to_collection":null,"budget":25000000,"genres":[{"id":18,"name":"Drama"},{"id":80,"name":"Crime"}],"homepage":"","id":278,"imdb_id":"tt0111161","original_language":"en","original_title":"The
@@ -61,60 +114,9 @@ http_interactions:
         at the Shawshank prison, where he puts his accounting skills to work for an
         amoral warden. During his long stretch in prison, Dufresne comes to be admired
         by the other inmates -- including an older prisoner named Red -- for his integrity
-        and unquenchable sense of hope.","popularity":81.472,"poster_path":"/hBcY0fE9pfXzvVaY4GKarweriG2.jpg","production_companies":[{"id":97,"logo_path":"/7znWcbDd4PcJzJUlJxYqAlPPykp.png","name":"Castle
+        and unquenchable sense of hope.","popularity":80.871,"poster_path":"/hBcY0fE9pfXzvVaY4GKarweriG2.jpg","production_companies":[{"id":97,"logo_path":"/7znWcbDd4PcJzJUlJxYqAlPPykp.png","name":"Castle
         Rock Entertainment","origin_country":"US"}],"production_countries":[{"iso_3166_1":"US","name":"United
         States of America"}],"release_date":"1994-09-23","revenue":28341469,"runtime":142,"spoken_languages":[{"english_name":"English","iso_639_1":"en","name":"English"}],"status":"Released","tagline":"Fear
-        can hold you prisoner. Hope can set you free.","title":"The Shawshank Redemption","video":false,"vote_average":8.7,"vote_count":22738}'
-  recorded_at: Thu, 01 Dec 2022 23:00:58 GMT
-- request:
-    method: get
-    uri: https://api.themoviedb.org/3/configuration?api_key=<tmdb_key>
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      User-Agent:
-      - Faraday v2.7.1
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Content-Type:
-      - application/json;charset=utf-8
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      Server:
-      - openresty
-      Date:
-      - Thu, 01 Dec 2022 22:59:11 GMT
-      Cache-Control:
-      - public, max-age=43200
-      Etag:
-      - W/"8ee6fd095d2be6e301251d766501a916"
-      Vary:
-      - Accept-Encoding
-      - Origin
-      X-Cache:
-      - Hit from cloudfront
-      Via:
-      - 1.1 87cff53a3b3c669d865b820d148e2d62.cloudfront.net (CloudFront)
-      X-Amz-Cf-Pop:
-      - HIO50-C2
-      Alt-Svc:
-      - h3=":443"; ma=86400
-      X-Amz-Cf-Id:
-      - J5L2cqULUfMjJJ_Bg08ZAhkj5B9CaDR4TOSqB3ipUD45EeeB0VFF5A==
-      Age:
-      - '107'
-    body:
-      encoding: ASCII-8BIT
-      string: '{"images":{"base_url":"http://image.tmdb.org/t/p/","secure_base_url":"https://image.tmdb.org/t/p/","backdrop_sizes":["w300","w780","w1280","original"],"logo_sizes":["w45","w92","w154","w185","w300","w500","original"],"poster_sizes":["w92","w154","w185","w342","w500","w780","original"],"profile_sizes":["w45","w185","h632","original"],"still_sizes":["w92","w185","w300","original"]},"change_keys":["adult","air_date","also_known_as","alternative_titles","biography","birthday","budget","cast","certifications","character_names","created_by","crew","deathday","episode","episode_number","episode_run_time","freebase_id","freebase_mid","general","genres","guest_stars","homepage","images","imdb_id","languages","name","network","origin_country","original_name","original_title","overview","parts","place_of_birth","plot_keywords","production_code","production_companies","production_countries","releases","revenue","runtime","season","season_number","season_regular","spoken_languages","status","tagline","title","translations","tvdb_id","tvrage_id","type","video","videos"]}'
-  recorded_at: Thu, 01 Dec 2022 23:00:58 GMT
+        can hold you prisoner. Hope can set you free.","title":"The Shawshank Redemption","video":false,"vote_average":8.7,"vote_count":22742}'
+  recorded_at: Fri, 02 Dec 2022 15:40:16 GMT
 recorded_with: VCR 6.1.0
